requirement.txt
torch
torchvision
Pillow
pyyaml